Fix coordinate scale pipeline, spawn safety, and legacy data migration

+++ server/BelterLife.Shared/Contracts/Api/SpawnResponse.cs
@@ -1,3 +1,3 @@
 namespace BelterLife.Shared.Contracts.Api;
 
-public record SpawnResponse(int SectorId, int ShipId, float SpawnX, float SpawnY, bool Repositioned = false);
+public record SpawnResponse(int SectorId, int ShipId, long SpawnX, long SpawnY, bool Repositioned = false);
