+++ server/BelterLife.Shared/Contracts/Hubs/AsteroidSnapshot.cs
@@ -1,3 +1,3 @@
 namespace BelterLife.Shared.Contracts.Hubs;
 
-public record AsteroidSnapshot(int AsteroidId, float X, float Y, float Radius, int VertexCount, float RotationOffset);
+public record AsteroidSnapshot(int AsteroidId, long X, long Y, float Radius, int VertexCount, float RotationOffset);
