+++ server/BelterLife.Shared/Entities/NpcStation.cs
@@ -4,7 +4,7 @@ public class NpcStation
 {
     public int Id { get; set; }
     public int SectorId { get; set; }
-    public float X { get; set; }
-    public float Y { get; set; }
+    public long X { get; set; }
+    public long Y { get; set; }
     public string Name { get; set; } = string.Empty;
 }
