+++ server/BelterLife.Shared/Entities/Ship.cs
@@ -5,8 +5,8 @@ public class Ship
     public int Id { get; set; }
     public string PlayerId { get; set; } = string.Empty;
     public int SectorId { get; set; }
-    public float X { get; set; }
-    public float Y { get; set; }
+    public long X { get; set; }
+    public long Y { get; set; }
     public float VelocityX { get; set; }
     public float VelocityY { get; set; }
     public float Heading { get; set; }
